Restrict accessible and deletable keys by user

+++ authorized_keys/views.py
@@ -1,4 +1,5 @@
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.core.exceptions import ValidationError
 from django.http import HttpResponse, HttpResponseRedirect
 from django.shortcuts import get_object_or_404, render
@@ -6,19 +7,22 @@ from django.urls import reverse
 
 from .models import PubSSHKey
 
+@login_required
 def index(request):
-	pubsshkey_list = PubSSHKey.objects.all()
+	mypk = request.user.pk
+	pubsshkey_list = PubSSHKey.objects.filter(user__pk=mypk)
 	context = {
 		'pubsshkey_list': pubsshkey_list,
 	}
 	return render(request, 'authorized_keys/index.html', context)
 
 
+@login_required
 def add(request):
 	try:
 		pubsshkey = PubSSHKey.create(
 				user_input=request.POST['content'],
-				# missing user
+				user=request.user,
 				title=request.POST.get('title'), # optional
 			)
 		pubsshkey.save()
@@ -33,13 +37,18 @@ def add(request):
 	return HttpResponseRedirect(reverse('authorized_keys:index'))
 
 
+@login_required
 def delete(request, key_id):
 	pubsshkey = get_object_or_404(PubSSHKey, pk=key_id)
 
-	# if k.user !=
-	#	error_message = "That's not your key"
-	# else:
-	pubsshkey.delete()
-	messages.success(request, 'Deleted key {}'.format(pubsshkey.fingerprint))
+	if pubsshkey.user.pk != request.user.pk:
+		messages.error(request, "That's not your key!")
+	else:
+		try:
+			pubsshkey.delete()
+		except Exception as e:
+			messages.error(request, "Cannot delete: {}".format(repr(e)))
+		else:
+			messages.success(request, 'Deleted key {}'.format(pubsshkey.fingerprint))
 
 	return HttpResponseRedirect(reverse('authorized_keys:index'))
